refactor(crypto): reorganize to canonical tier structure
