+++ server/src/openapi.rs
@@ -239,13 +239,13 @@ use utoipa::{Modify, OpenApi};
         crate::chat::screenshare::start,
         crate::chat::screenshare::stop,
         // Crypto
-        crate::crypto::handlers::upload_keys,
-        crate::crypto::handlers::get_backup,
-        crate::crypto::handlers::upload_backup,
-        crate::crypto::handlers::get_backup_status,
-        crate::crypto::handlers::get_own_devices,
-        crate::crypto::handlers::get_user_keys,
-        crate::crypto::handlers::claim_prekey,
+        crate::crypto::upload_keys,
+        crate::crypto::get_backup,
+        crate::crypto::upload_backup,
+        crate::crypto::get_backup_status,
+        crate::crypto::get_own_devices,
+        crate::crypto::get_user_keys,
+        crate::crypto::claim_prekey,
         // Bots
         crate::api::bots::list_applications,
         crate::api::bots::create_application,
